增加bakleader对元数据同步状态的判断。便于竞选新leader

+++ easyTask-X-core/src/main/java/com/github/liuche51/easyTaskX/dto/RegBroker.java
@@ -42,7 +42,7 @@ public class RegBroker extends RegNode {
         super(host, port);
     }
 
-    public RegBroker(String host, int port, Short dataStatus) {
+    public RegBroker(String host, int port, Integer dataStatus) {
         super(host, port, dataStatus);
     }
